feat: on failure, redefine stack datalist if null
if there's a request failure, the user invokes the stackPage method with
the datalist paramater as null. Stack then redifines the datalist to be
a list of predefined preload objects.

This ensures a continuous userexperience in the event of a request
failure. It also allows Stack to keep track of the count of items in
each block and how to properly dispose adapter contents upon block
removal.